FIx for thermal vision being weird
